amend the 2022/12/28 question update InventorGraph
