+++ app/user-agent/apis/package.go
@@ -2,12 +2,8 @@ package apis
 
 import (
 	"github.com/gin-gonic/gin/binding"
-	amodels "go-admin/app/admin/models"
-	aservice "go-admin/app/admin/service"
-	adto "go-admin/app/admin/service/dto"
 	"go-admin/app/user-agent/models"
 	"net/http"
-	"sort"
 	"strconv"
 
 	"github.com/gin-gonic/gin"
@@ -468,19 +464,15 @@ func (e Package) DeletePackagePatent(c *gin.Context) {
 
 //---------------------------------------------------patent--graph-------------------------------------------------------
 
-// GetTheGraphByPackageId3
+// GetRelationGraphByPackage
 // @Summary 获取专利包中专利的发明人的关系
 // @Description  获取专利包中专利的发明人的关系
 // @Tags 专利表
 // @Router /api/v1/user-agent/package/{packageId}/relationship3 [get]
 // @Security Bearer
-func (e Package) GetTheGraphByPackageId3(c *gin.Context) {
-	sup := service.UserPatent{}
-	su := aservice.SysUser{}
+func (e Package) GetRelationGraphByPackage(c *gin.Context) {
 	sp := service.Patent{}
-	//gservice := service.Node{}
 	reqp := dto.PatentsIds{}
-	requ := adto.SysUserById{}
 	spp := service.PatentPackage{}
 	reqpp := dto.PackagePageGetReq{}
 	Inventorgraph := models.Graph{}
@@ -504,305 +496,20 @@ func (e Package) GetTheGraphByPackageId3(c *gin.Context) {
 		reqp.PatentIds[i] = listpp[i].PatentId
 	}
 	listp := make([]models.Patent, 0)
-	var count2 int64
 	err = e.MakeContext(c).
 		MakeOrm().
 		MakeService(&sp.Service).
 		Errors
-	err = sp.GetPageByIds(&reqp, &listp, &count2)
-	links := make([]models.Link, 0)
 
-	err = e.MakeContext(c).
-		MakeOrm().
-		MakeService(&sup.Service).
-		Errors
-	listInventorId := make(map[string]int)
-	listup2 := make([]models.InventorPatent, 0)
-	FindTheInventorFromPatents(&listInventorId, &listup2, listp)
-	usertimes := make(map[int]int)
-	for i := 0; i < len(listup2); i++ {
-		if usertimes[listup2[i].UserId] == 0 {
-			usertimes[listup2[i].UserId] = 1
-		} else {
-			usertimes[listup2[i].UserId]++
-		}
-
+	err = sp.GetPageByIds(&reqp, &listp, &count)
+	if err != nil {
+		e.Logger.Error(err)
+		return
 	}
-	usertimes1 := rankByWordCount(usertimes)
-	var members int
-	if len(usertimes1) < 500 {
-		members = len(usertimes1)
-	} else {
-		members = 500
+	err = sp.GetGraphByPatents(listp, &Inventorgraph)
+	if err != nil {
+		e.Logger.Error(err)
+		return
 	}
-	UserIsNode := make([]bool, members)
-	var StrongRelationNode int
-	if members >= 100 {
-		StrongRelationNode = 10
-	} else {
-		StrongRelationNode = members / 10
-	}
-	NodeList := make([]models.Node, StrongRelationNode)
-	for i := 0; i < StrongRelationNode; i++ {
-		NodeList[i].NodeCategory = i
-		NodeList[i].NodeId = strconv.FormatInt(int64(usertimes1[i].Key), 10)
-		NodeList[i].NodeValue = usertimes1[i].Value
-		UserIsNode[i] = true
-	}
-	userspatents := make([]models.OneUserPatents, members)
-	for i := 0; i < len(listup2); i++ {
-		for j := 0; j < members; j++ {
-			if listup2[i].UserId == usertimes1[j].Key {
-				userspatents[j].Patentsid = append(userspatents[j].Patentsid, listup2[i].PatentId)
-				break
-			}
-		}
-	}
-	useruserrelation1 := make(map[int]int)
-	first10 := StrongRelationNode
-	firstlinks := first10
-	for i := 0; i < first10; i++ {
-		for j := i + 1; j < first10; j++ {
-			RelationExist := 0
-			for z := 0; z < len(userspatents[i].Patentsid); z++ {
-				for z1 := 0; z1 < len(userspatents[j].Patentsid); z1++ {
-					if userspatents[i].Patentsid[z] == userspatents[j].Patentsid[z1] {
-						RelationExist++
-						break
-					}
-				}
-			}
-			if RelationExist != 0 {
-				useruserrelation1[i*10+j] = RelationExist
-			}
-		}
-	}
-	useruserrelation2 := rankByWordCount(useruserrelation1) //给边排序
-	for i := 0; i < minresult(firstlinks, len(useruserrelation2)); i++ {
-		var nowlink models.Link
-		nowlink.Source = strconv.FormatInt(int64(usertimes1[useruserrelation2[i].Key/10].Key), 10)
-		nowlink.Target = strconv.FormatInt(int64(usertimes1[useruserrelation2[i].Key%10].Key), 10)
-		nowlink.Value = useruserrelation2[i].Value
-		links = append(links, nowlink)
-	}
-
-	if len(NodeList) < 1 {
-		Inventorgraph.Links = links
-		Inventorgraph.Nodes = NodeList
-		e.GraphOK(NodeList, links, "查询成功")
-	}
-	useruserrelation3 := make(map[int]int)
-	secondLinks := 200
-	ExtendNodeTime := make([]int, members)
-
-	for i := 0; i < first10; i++ {
-		for j := first10; j < members; j++ {
-			RelationExist := 0
-			for z := 0; z < len(userspatents[i].Patentsid); z++ {
-				for z1 := 0; z1 < len(userspatents[j].Patentsid); z1++ {
-					if userspatents[i].Patentsid[z] == userspatents[j].Patentsid[z1] {
-						RelationExist++
-						break
-					}
-				}
-			}
-			if RelationExist != 0 {
-				useruserrelation3[i*500+j] = RelationExist
-			}
-		}
-	}
-	useruserrelation4 := rankByWordCount(useruserrelation3)
-	NodelistIdToTimeList := make(map[string]int)
-
-	for i := 0; i < minresult(secondLinks, len(useruserrelation4)); i++ {
-		source := useruserrelation4[i].Key / 500
-		target := useruserrelation4[i].Key % 500
-		if ExtendNodeTime[source] >= 5 {
-			continue
-		} else {
-			if UserIsNode[target] == false {
-				UserIsNode[target] = true
-				ExtendNodeTime[source]++
-				var nowlink models.Link
-				var nowNode models.Node
-				nowNode.NodeCategory = NodeList[source].NodeCategory
-				nowNode.NodeId = strconv.FormatInt(int64(usertimes1[target].Key), 10)
-				NodelistIdToTimeList[nowNode.NodeId] = target
-				nowlink.Source = strconv.FormatInt(int64(usertimes1[source].Key), 10)
-				nowlink.Target = strconv.FormatInt(int64(usertimes1[target].Key), 10)
-				nowlink.Value = useruserrelation4[i].Value
-				links = append(links, nowlink)
-				NodeList = append(NodeList, nowNode)
-			}
-		}
-	}
-
-	useruserrelation5 := make(map[int]int)
-
-	for i := first10; i < len(NodeList); i++ {
-		iToUserspatentsPosition := NodelistIdToTimeList[NodeList[i].NodeId]
-		for j := i + 1; j < len(NodeList); j++ {
-			RelationExist := 0
-			jToUserspatentsPosition := NodelistIdToTimeList[NodeList[j].NodeId]
-			for z := 0; z < len(userspatents[iToUserspatentsPosition].Patentsid); z++ {
-				for z1 := 0; z1 < len(userspatents[jToUserspatentsPosition].Patentsid); z1++ {
-
-					if userspatents[iToUserspatentsPosition].Patentsid[z] == userspatents[jToUserspatentsPosition].Patentsid[z1] {
-						RelationExist++
-						break
-					}
-				}
-			}
-			if RelationExist != 0 {
-				useruserrelation5[i*500+j] = RelationExist
-			}
-
-		}
-	}
-	useruserrelation6 := rankByWordCount(useruserrelation5)
-	thirdlinks := 50
-	for i := 0; i < minresult(thirdlinks, len(useruserrelation6)); i++ {
-		source := useruserrelation6[i].Key / 500
-		target := useruserrelation6[i].Key % 500
-
-		if ExtendNodeTime[source] >= 2 {
-			continue
-		} else {
-			ExtendNodeTime[source]++
-			var nowlink models.Link
-			nowlink.Source = NodeList[source].NodeId
-			nowlink.Target = NodeList[target].NodeId
-			nowlink.Value = useruserrelation6[i].Value
-			links = append(links, nowlink)
-		}
-	}
-
-	listu := make([]amodels.SysUser, 0)
-	err = e.MakeContext(c).
-		MakeOrm().
-		Bind(&requ, nil).
-		MakeService(&su.Service).
-		Errors
-	for i := 0; i < len(NodeList); i++ {
-		requ.Id, err = strconv.Atoi(NodeList[i].NodeId)
-		var user1 amodels.SysUser
-		user1.UserId = requ.Id
-		for k, v := range listInventorId {
-			if v == requ.Id {
-				user1.Username = k
-			}
-		}
-		listu = append(listu, user1)
-	}
-	//err = e.MakeContext(c).
-	//	MakeOrm().
-	//	MakeService(&gservice.Service).
-	//	Errors
-	max := 0
-	min := 100000
-	NodeList[0].NodeValue = usertimes1[0].Value
-	NodeList[0].NodeSymbolizeSize = 50
-	NodeList[0].NodeName = listu[0].Username
-	for i := 1; i < first10; i++ {
-		NodeList[i].NodeValue = usertimes1[i].Value
-		if NodeList[i].NodeValue > max {
-			max = NodeList[i].NodeValue
-		}
-		if NodeList[i].NodeValue < min {
-			min = NodeList[i].NodeValue
-		}
-	}
-	for i := first10; i < len(NodeList); i++ {
-		NodeList[i].NodeValue = usertimes1[NodelistIdToTimeList[NodeList[i].NodeId]].Value
-		if NodeList[i].NodeValue > max {
-			max = NodeList[i].NodeValue
-		}
-		if NodeList[i].NodeValue < min {
-			min = NodeList[i].NodeValue
-		}
-	}
-
-	for i := 1; i < len(NodeList); i++ {
-		NodeList[i].NodeSymbolizeSize = float32(float32(NodeList[i].NodeValue*30) / float32(maxresult((max), 1)))
-		NodeList[i].NodeName = listu[i].Username
-	}
-	Inventorgraph.Links = links
-	Inventorgraph.Nodes = NodeList
 	e.OK(Inventorgraph, "查询成功")
-}
-
-// FindTheInventorFromPatents --------------------------------------------------------------------------
-// 查找patents中的发明人
-func FindTheInventorFromPatents(listInventorId *map[string]int, listup2 *[]models.InventorPatent, listp []models.Patent) error {
-	var err error
-	count := 0
-	for z := 0; z < len(listp); z++ {
-		words := make([]string, 0)
-		for i := 0; i < len(listp[z].PatentProperties); i++ {
-			if listp[z].PatentProperties[i] == '"' && listp[z].PatentProperties[i+1] == 'P' && listp[z].PatentProperties[i+2] == 'I' && listp[z].PatentProperties[i-1] == ',' {
-				now := i + 8
-				for j := i + 8; j < len(listp[z].PatentProperties); j++ {
-					if listp[z].PatentProperties[j] == '"' {
-						words = append(words, listp[z].PatentProperties[now:j])
-						break
-					}
-					if listp[z].PatentProperties[j] == ';' {
-						words = append(words, listp[z].PatentProperties[now:j])
-						now = j + 1
-					}
-				}
-				break
-			}
-		}
-		for i := 0; i < len(words); i++ {
-			_, ok := (*listInventorId)[words[i]]
-			if !ok {
-				(*listInventorId)[words[i]] = count
-				count++
-			}
-			var inventorpatent models.InventorPatent
-			inventorpatent.PatentId = listp[z].PatentId
-			inventorpatent.UserId = (*listInventorId)[words[i]]
-			*listup2 = append(*listup2, inventorpatent)
-		}
-	}
-	return err
-}
-
-// --------------------------------------------------------------------------------------------------------------------
-// map按value的值排序
-func rankByWordCount(wordFrequencies map[int]int) PairList {
-	pl := make(PairList, len(wordFrequencies))
-	i := 0
-	for k, v := range wordFrequencies {
-		pl[i] = Pair{k, v}
-		i++
-	}
-	sort.Sort(sort.Reverse(pl))
-	return pl
-}
-
-type Pair struct {
-	Key   int
-	Value int
-}
-
-type PairList []Pair
-
-func (p PairList) Len() int           { return len(p) }
-func (p PairList) Less(i, j int) bool { return p[i].Value < p[j].Value }
-func (p PairList) Swap(i, j int)      { p[i], p[j] = p[j], p[i] }
-func minresult(a1 int, a2 int) int {
-	if a1 >= a2 {
-		return a2
-	} else {
-		return a1
-	}
-}
-func maxresult(a1 int, a2 int) int {
-	if a1 >= a2 {
-		return a1
-	} else {
-		return a2
-	}
 }
