+++ app/user-agent/router/package.go
@@ -22,10 +22,10 @@ func registerPackageRouter(v1 *gin.RouterGroup, authMiddleware *jwt.GinJWTMiddle
 		r.PUT("/:id", api.Update)
 		r.DELETE("/:id", api.Delete)
 
-		r.GET("/:id/patent", api.GetPackagePatents)              //显示专利包内专利√
-		r.POST("/:id/patent", api.InsertPackagePatent)           //将专利加入专利包√
-		r.DELETE("/:id/patent/:pid", api.DeletePackagePatent)    //取消加入该专利包
-		r.GET("/:id/patent/:PNM/isExist", api.IsPatentInPackage) //取消加入该专利包
-		r.GET("/:id/relationship3", api.GetTheGraphByPackageId3) //获取该专利包的发明人关系图谱(点位置随机、大小数量和线的数量、值根据数据生成)
+		r.GET("/:id/patent", api.GetPackagePatents)                //显示专利包内专利√
+		r.POST("/:id/patent", api.InsertPackagePatent)             //将专利加入专利包√
+		r.DELETE("/:id/patent/:pid", api.DeletePackagePatent)      //取消加入该专利包
+		r.GET("/:id/patent/:PNM/isExist", api.IsPatentInPackage)   //取消加入该专利包
+		r.GET("/:id/relationship3", api.GetRelationGraphByPackage) //获取该专利包的发明人关系图谱(点位置随机、大小数量和线的数量、值根据数据生成)
 	}
 }
